+++ src/finlang.test/TranspilerTest/ExSln2Fixture.cs
@@ -1,4 +1,5 @@
-﻿using Xunit.Sdk;
+﻿using finlang.Transpiler;
+using Xunit.Sdk;
 
 namespace finlang.TranspilerTest;